+++ Package.swift
@@ -5,6 +5,7 @@ import PackageDescription
 
 let package = Package(
     name: "ZoomImageViewer",
+    defaultLocalization: "en",
     platforms: [
         .iOS(.v15)
     ],
@@ -19,7 +20,10 @@ let package = Package(
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
-            name: "ZoomImageViewer"
+            name: "ZoomImageViewer",
+            resources: [
+                .process("Localizable.xcstrings")
+            ]
         ),
         .testTarget(
             name: "ZoomImageViewerTests",